Add football Schedule class
The Schedule class can be used to iterate over a team's schedule for the
season and includes high-level information for each game. The schedule
can be queried by supplying either the squad ID or the full team name.
Every match will be an instance of the Game class in the same module.

Signed-Off-By: Robert Clark <robdclark@outlook.com>