Add support for football (soccer)
Now that sports-reference.com has a mature website for football, it
should be added to this package to allow easy queries of player, team,
and league statistics and information using a Python API. This is only
the initial release of football metrics for sportsreference as future
updates are planned as well to further enhance the package.

In addition to the addition of the new section, the interface for
connecting to teams uses a new format, diverging from other sections in
the package to allow teams to more easily and directly be queried. This
new format doesn't force users to first pull a master list of teams,
then pull a specific team, but instead uses a lookup-table to search for
a specific team to make it easier. The hope is that this simpler method
will be used in other sections for further ease of use and be included
prior to the official 1.0 release of sportsreference.

Signed-Off-By: Robert Clark <robdclark@outlook.com>